% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGB.R
\name{getGB}
\alias{getGB}
\title{This is an internal function in AnnotationBustR that is amodified version of ape's read.GenBank () to get accession numbers and info from GenBank}
\usage{
getGB(access.nb, seq.names = access.nb, species.names = TRUE,
  gene.names = FALSE, as.character = FALSE)
}
\arguments{
\item{access.nb}{A vector of GenBank accession numbers.}

\item{seq.names}{Names for sequences}

\item{species.names}{Logical. Should species names be returned.Default is TRUE.}

\item{gene.names}{Logical. Should gene names be returned.Default is FALSE.}

\item{as.character}{Logical. Should returned object strings be treated as character. Default is FALSE.}
}
\value{
A fasta with sequence names and length info
}
\description{
This is an internal function in AnnotationBustR that is amodified version of ape's read.GenBank () to get accession numbers and info from GenBank
}
\details{
internal function returns accessions and fasta files to measure length in AnnotationBustR
}
\references{
Popescu, Andrei-Alin, Katharina T. Huber, and Emmanuel Paradis. "ape 3.0: New tools for distance-based phylogenetics and evolutionary analysis in R." Bioinformatics 28.11 (2012): 1536-1537.
}
\keyword{internal}
