% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotPos.R
\name{GetSeqPos}
\alias{GetSeqPos}
\title{Finds annotation positions based on search terms that can later be used to seperate sequences into their annotated components.}
\usage{
GetSeqPos(accessions, genes, duplicate.genes = c("tRNA_Ser", "tRNA_Leu"))
}
\arguments{
\item{accessions}{A vector of GenBank accession numbers.}

\item{genes}{A data frame of search terms. Pre-compiled search term lists are available as data with this package for mitogenomes and rDNA.}

\item{duplicate.genes}{A character vector containing duplicate gene names found in the annotation. Ex. serine and leucine in mitogenomes.}
}
\value{
A table of start and stop positions of class annotPos for all the genes specified for all accession numbers that can be used to bust sequences using AnnotationBustR. Negative numbers indicate the sequence is a complement
}
\description{
Finds annotation positions based on search terms that can later be used to seperate sequences into their annotated components.
}
\examples{
ncbi.accessions<-c("FJ706343","FJ706292")#vector of two NCBI accession numbers to get the annotation positions of.
data(rDNA.Genes)#load rDNA search terms
my.seq.pos<-GetSeqPos(ncbi.accessions, rDNA.Genes, duplicate.genes= NULL)#Get rDNA gene positions for each sequence. There are no gene duplicates.
}

