% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{mtDNAtermsPlants}
\alias{mtDNAtermsPlants}
\title{Mitochondrial DNA Search Terms for Plants}
\format{A data frame of of 248 rows and 3 columns
\itemize{
  \item Locus: Locus name, FASTA files will be written with this name
  \item Type: Type of subsequence, either CDS,tRNA,rRNA,misc_RNA, or D-loop
  \item Name:Name of synonym for a locus to search for
}}
\usage{
mtDNAtermsPlants
}
\description{
A data frame containing search terms for plant mitochondrial loci. Can be subset for loci of 
 interest. Columns are as follows and users should follow the column format if they wish to
 add search terms using the MergeSearchTerms function:
}
\seealso{
\code{\link{MergeSearchTerms}}
}
\keyword{datasets}
