% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{rDNAterms}
\alias{rDNAterms}
\title{Ribosomal DNA (rDNA) Search Terms}
\format{
A data frame of of 14 rows and 3 columns
\itemize{
  \item Feature: Feature name, FASTA files will be written with this name.
  \item Type: Type of feature, either rRNA or misc_RNA.
  \item Name: Name of synonym for a feature to search for.
}
}
\usage{
rDNAterms
}
\description{
A data frame containing search terms for ribosomal RNA loci. Can be subset for loci of 
 interest. Columns are as follows and users should follow the column format if they wish to
 add search terms using the MergeSearchTerms function:
}
\seealso{
\code{\link{MergeSearchTerms}}
}
\keyword{datasets}
