% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationBustR.R
\name{AnnotationBust}
\alias{AnnotationBust}
\title{Breaks up genbank sequences into their annotated components based on positions found using the get.seq.pos function.}
\usage{
AnnotationBust(Seq.Positions)
}
\arguments{
\item{Seq.Positions}{Object of class Annot.Pos with the starting and stopping position for all genes.}
}
\value{
Returns a fasta file for each gene in the object of class Annot.Pos supplied into the function.
}
\description{
Breaks up genbank sequences into their annotated components based on positions found using the get.seq.pos function.
}

