% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLongestSeq.R
\name{FindLongestSeq}
\alias{FindLongestSeq}
\title{Find the longest sequence for each species from a list of GenBank accession numbers.}
\usage{
FindLongestSeq(accessions)
}
\arguments{
\item{accessions}{A vector of GenBank accession numbers.}
}
\value{
A list of genbank accessions numbers for the longest sequence for each taxon in a list of accession numbers.
}
\description{
Find the longest sequence for each species from a list of GenBank accession numbers.
}
\examples{
genbank.accessions<-c("KP978059.1","KP978060.1","JX516105.1","JX516111.1")#vector of 4 genbank accessions, two each for two species
long.seq.result<-FindLongestSeq(genbank.accessions)#returns the longest sequence respectively for the two species.
}

