% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSearchTerms.R
\name{MergeSearchTerms}
\alias{MergeSearchTerms}
\title{Allows for the merging of two tables containing search terms to expand search term database for getAnnotPos.}
\usage{
MergeSearchTerms(..., sort.gene = FALSE)
}
\arguments{
\item{...}{the data frames of search terms you want to combine into a single list. The Data frames should have stringsAsFactors=FALSE if you want to sort them.}

\item{sort.gene}{Should the final data frame be sorted by gene name? Default is FALSE.}
}
\value{
A new merged data frame with all the search terms combined from the lists supplied. If sort.gene=TRUE, genes will be sorted by name.
}
\description{
Allows for the merging of two tables containing search terms to expand search term database for getAnnotPos.
}
\examples{
data(MitoGenes) #load the list of search terms for mitochondrial genes
add.name<-data.frame("D_loop","D-loop","note", "d-loop")#Create a data frame with a new search term.
colnames(add.names)<-colnames(Mito.Genes)# make the column names the same for combination.
MergeSearchTerms(add.name, Mito.Genes, sort.gene=FALSE)#Run the merge search term function without sorting based on gene name.
MergeSearchTerms(add.name, Mito.Genes, sort.gene=TRUE)#Run the merge search term function with sorting based on gene name.
}

